Change gargabe collection for the LISP interface.
add to class JIT-STATE the flag CLEAR-STATE-MANUALLY, which means, if non-NIL, to not call #'FN-JIT-CLEAR-STATE on garbage collection.
add to #'NEW-STATE the flag CLEAR-STATE-MANUALLY.
add to #'NEW-STATE calling normal (respectively full) garbage collection automatically every *NEW-STATE-MAX-CALLS-BEFORE-GC* (respectively *NEW-STATE-MAX-CALLS-BEFORE-FULL-GC*) calls to it, which makes it usable in some LISPs which do not call garbage collection often enough.
add to macro WITH-NEW-STATE the flag CLEAR-STATE-MANUALLY.
remove macro IMPERATIVE-BIND, which was just a LET-form in disguise.
remove macro WITH-JIT.
adapt examples.